Refactor repository for better structure and clarity
This commit introduces several changes to improve the structure and maintainability of the repository.

The main changes are:
- Simplified the `clintrials/dosefinding` module structure by moving the logic from `crm/design.py` and `efftox/design.py` into their parent modules.
- Removed redundant and unnecessary files from the root directory, including duplicate `README` and `LICENSE` files, and obsolete Python scripts (`imp.py`, `sitecustomize.py`).
- Moved the `tutorials` directory into the `docs` directory to consolidate all documentation-related materials.
- Updated all references to the moved `tutorials` directory.
