--- clintrials/dosefinding/crm/__init__.py
@@ -1,3 +1,0 @@
-from .design import CRM, crm, crm_dtp_detail
-
-__all__ = ["CRM", "crm", "crm_dtp_detail"]
