--- clintrials/dosefinding/efftox/__init__.py
@@ -1,15 +1,0 @@
-from .design import (
-    EffTox,
-    InverseQuadraticCurve,
-    LpNormCurve,
-    efftox_dtp_detail,
-    solve_metrizable_efftox_scenario,
-)
-
-__all__ = [
-    "EffTox",
-    "LpNormCurve",
-    "InverseQuadraticCurve",
-    "efftox_dtp_detail",
-    "solve_metrizable_efftox_scenario",
-]
