+++ docs/user_guide/tutorials.md
@@ -1,5 +1,5 @@
 # Tutorials
 
-This section contains introductory material to help you use clintrials. The notebooks in `tutorials/` demonstrate dose finding and other analyses. You can run them interactively after installing the optional dependencies.
+This section contains introductory material to help you use clintrials. The notebooks in `docs/tutorials/` demonstrate dose finding and other analyses. You can run them interactively after installing the optional dependencies.
 
 More examples will be added over time.
