--- sitecustomize.py
@@ -1,4 +1,0 @@
-import unittest.runner
-
-if not hasattr(unittest.runner, "_TextTestResult"):
-    unittest.runner._TextTestResult = unittest.runner.TextTestResult
